// src/common/enums/leaveStatus.enum.ts
export enum LeaveStatus {
  Pending_HR = 'Pending HR Review',
  Pending_ProgramHead = 'Pending to Program Head',
  Pending_Admin = 'Pending Admin Review',
  Rejected_Request ='Archive for Rejected Request',
  Approved_Request = 'Archive for Approved Request'
}

export enum ResultStatus {
  Approved = 'Approved',
  Rejected = 'Rejected',
  Pending = 'Pending',
}